[swss.service]: move redis db flushall from syncd to swss
When we restart swss, syncd is also restarted. currently the db flushall
is in syncd, depends on timing, swss may already put some new information into the
db and then db get flushed, causing swss init process to fail.

The original reason to put flushall in syncd is to support syncd only restart,
since it is not the most common cause, move the flushall to swss.