[interface_facts] improve interface facts checking performance
Allow callers of interface_facts to pass in a list of ports of that
are expected to be up. The script will check their link status, return
list of ports that has link went down unexpectedly.

The caller, instead of looping in ansible to check all interface status,
can simply check if the link down list is empty.