[test case] Add test: restart swss service
There was a script to restart swss, it was used as utility by a
few other tests. And it didn't wait long enough for swss service
to recover, it didn't check if the system recovered eventually.

This test also restarts dhcp_relay service to make sure that it
will listen on right interfaces. So the system would be left at
a healthy state.