[test infrastructure] allow overriding recover system actioin (#480)
Automatical recover test environment is great to let test to run in a healthy environment.

There are scenarios that running a test was to repeat an issue. In this case, we want the
test to fail when prolem happens and leave the system in the state for investigation.

Adding opiton to prevent test to recover system. Default behavior is still recoverying
unhealthy system.