[everflow]: Add --table_name argument in EVERFLOW test
acl-loader will support --table_name option so that only specific table
will be updated. This test covers the scenario that we have data plane
ACLs while we want to apply and remove EVERFLOW ACLs. The 14 data plane
ACLs are applied first and we check if they stay there during the whole
EVERFLOW test. These ACLs will be removed at the end of the test.

Signed-off-by: Shuotian Cheng <stcheng_89@hotmail.com>
Signed-off-by: Shu0T1an ChenG <shuche@microsoft.com>