Merged PR 882147: Swap N9232 nightly test device roles
Swap because T0 device runs fast-reboot test, which requires logging into the fanout switch(es). However, str-n9232-acs-1 is connected to a fanout switch which we are currently test-running SONiC on (str-7060cx-32s-08). Unfortunately, the credentials are different, and Ansible fails to SSH to this switch, so the test consistently would fail. This PR swaps the roles of the devices so that the T1 test N9232 device connects to the SONiC fanout, and now the T0 device connects only to fanout switches running Arista EOS.