[fast-reboot] don't restore original image when no new image defined
It appears taht ansible ignores 'when' condition for tasks under 'always' section.
When new image is not defined. The original image is not set. So when test executed
to always section trying to restore the origin image, an ansible error occurs.

Use notify/handler mechanism to make sure original image is restored when necessary,
even under failure conditions.

Signed-off-by: Ying Xie <ying.xie@microsoft.com>