[start vm] provide a method to respin some vms
Define respin_vm to override the test for VM existence.
Any VM defined in the list will be destroyed and restarted.
This provides a manual method to respin some VMs when they
are down.

I encountered VM down when deploying topology. Without a
method to respin these VMs, the other way is to reboot/cleanup
server and restart all VMs. The later method disrupts all
topologies on the server.

After respining individual VMs, the affected topology needs to
be removed and deployed again.

Signed-off-by: Ying Xie <ying.xie@microsoft.com>