[logAnalyzer] Avoid logAnalyzer logs being overwritten
In logAnalyzer related scripts, ansible_date_time is used for generating
testname_unique which will be used as folder name for storing
logAnalyzer results. The drawback is that ansible_date_time is fixed
in single ansible playbook execution. If logAnalyzer is called
multiple times, the earlier results could be overwritten by the latest
results.

The fix is to replace ansible_date_time with:
{{lookup('pipe','date +%Y-%m-%d-%H:%M:%S')}}

There are two differences:
1. Each time the lookup plugin is called, the returned date and time
   would be different (interval of two calls longer than 1 second)
2. The lookup plugin gets date and time of the sonic-mgmt container,
   not date and time of DUT.

Signed-off-by: Xin Wang <xinw@mellanox.com>