[fast/warm reboot] add knob to clean up old iamges on DUT before warm/fast reboot
When new image is specified for fast/warm reboot. The new image will be installed.
However, if the specified image is already installed on the target DUT, then
sonic_install will fail and fast/warm reboot will reboot into current image.

Add a knob to cleanup old images so that the installing of new image will have a
better chance to succeed.

Signed-off-by: Ying Xie <ying.xie@microsoft.com>