[fast/warm reboot] improve new image installation code (#877)
* [fast/warm reboot] improve new image installation code

- Allow new_sonic_image being defined as empty string. It causes skipping image installation.
- Rename new_image_location to a generic name.
- Display defined new image url.

Signed-off-by: Ying Xie <ying.xie@microsoft.com>

* [fast/warm reboot] allow DUT to stay in the warm/fast reboot target release

This feature is needed in order to test ugprade path. Where we might upgrade from one version
to another, and more. We want the system to stay in target release for next steps.

Signed-off-by: Ying Xie <ying.xie@microsoft.com>

* Address review comments, test issues and some minor touch-ups

Signed-off-by: Ying Xie <ying.xie@microsoft.com>

* [fast/warm reboot] add knob to clean up old iamges on DUT before warm/fast reboot

When new image is specified for fast/warm reboot. The new image will be installed.
However, if the specified image is already installed on the target DUT, then
sonic_install will fail and fast/warm reboot will reboot into current image.

Add a knob to cleanup old images so that the installing of new image will have a
better chance to succeed.

Signed-off-by: Ying Xie <ying.xie@microsoft.com>

* address review issue