[warm-reboot] Update the expected route for ipv6 loopback interface
After quagga is replaced with frr on SONiC, the route for IPv6 loopback
interface advertised to peer has been changed from prefix to the exact
IPv6 address of the interface. The automation scripts need to be updated
accordingly, otherwise the arista.py script would stuck in a while loop
checking IP routes.

Signed-off-by: Xin Wang <xinw@mellanox.com>