[vm_set] Improve the start-vms performance
The original approach starts and configures the VMs sequentially.
It takes more than 3 hours to start 32 virtual machines. This
change is to start all the VMs, then configure them one by one.
With this change, starting 32 VMs needs around 30-40 minutes.

Another change in this commit is to configure 'autostart' for
VMs so that the VMs will automatically start running after host
server is rebooted.

Signed-off-by: Xin Wang <xinw@mellanox.com>