[loganalyzer] Fix the files not copied issue in run_command_with_log_analyzer.yml (#963)
The copy files task was after the fail tests. In case of failure, the
copy task would never get a chance to run. This commit
adjusted the task sequence. In case of failure, copy the files, then
fail the test.

The original copy task copies files with deep folder structure.
This issue was also fixed in this commit.

Signed-off-by: Xin Wang <xinw@mellanox.com>