[loganalyzer] Generate dump of log within 1 hour by default if failed
By default the log analyzer generate a dump which collect all the
available log files by default in case of failure. This unnecessary and
the dump file could be too big.
This fix is to generate a dump to collect log within 1 hour by default.
If more log is needed, parameter 'dump_since' can be used.

Signed-off-by: Xin Wang <xinw@mellanox.com>