[warm reboot] save config after upgrading into new image (#1003)
* [warm/fast reboot] make sure that /etc/sonic/config_db.json exsits after upgrade

Signed-off-by: Ying Xie <ying.xie@microsoft.com>

* [warm reboot] save config after warm reboot into new image

When new image is defined, test removed /host/config_db.json
before warm rebooting. So after the device boots up, it will
miss /etc/sonic/config_db.json. It is not an issue for the
device to stay up. But it will be an issue when device reboot
again (cold or fast).

Signed-off-by: Ying Xie <ying.xie@microsoft.com>

* review comments