[sonic_upgrade] remove none-current and non-next images (#1020)
As the hard drive permits, we keep a few history images in the past so
that we could easily go back to them.

Recent test failure shows a downside of that decision. When a test failed
and leave an installed image in broken state. We likely restore the system
by booting into another working image. However, if the broken image is not
removed before installation happens again, because the target image exists,
the installation could be skipped or not fixing the existing issue. So
when we boot into the image again, the device is still in broken state.

Removing all non-current and non-next images give the DUT a better chance
to start a clean test.

Signed-off-by: Ying Xie <ying.xie@microsoft.com>