[loganalyzer] Fix the match_file_list accumulating issue
The match_file_list is not initialized to empty list when it is used.
The side effect is that when the log analyzer is called multiple times,
test match file used in earlier analysis will be used in later analysis
and cause issue. The fix is to initialize match_file_list to empty list
before using it.

Signed-off-by: Xin Wang <xinw@mellanox.com>