[Config Test] swss exit after remove port channel due to SAI not found intf (#1028)
When executing the config test case, it has the possibility to cause swss exit. It is due to the intfOrch may be slower than portOrch

When the issue not happens, the timeline would be

1. Playbook remove IP from port channel
intfOrch want to remove l3 intf, but neighOrch still have reference and the intfOrch will wait to next run

2. neighOrch decrease the reference count
neighOrch finish the reference count decrement and then intfOrch remove l3 intf to SAI

3. Playbook remove port channel
portOrch remove this port in SAI

When the issue happens, the timeline would be

1. Playbook remove IP from port channel
intfOrch want to remove l3 intf, but neighOrch still have reference and the intfOrch will wait to next run

2. Playbook remove port channel
portOrch remove this port and its l3 intf in SAI

3. neighOrch decrease the reference count
neighOrch finish the reference count decrement and then intfOrch remove l3 intf to SAI whcih was removed by step 2. Then the swss will exception