Fix the exit code return issue in shell scripts (#1107)
* Improve exit code return in shell scripts

In shell script, by default exit code of the last command
in commands connected by pipe is returned. If previous
command returned none zero exit code, the error is
not captured. If ansible run such script in playbook, the
script issue is ignored and ansible is not aware of it.

The fix is to add "set -o pipefail" in shell scripts. And
put the commands embedded in for loop statement
to a separate statement.

Signed-off-by: Xin Wang <xinw@mellanox.com>

* Restore code that was overwritten by copy and paste