[fdb] More strict checking of FDB entry count
In the FdbConfigReloadTest, we observed that the SDK
FDB count and number of MACs in "show mac" output
may be 0. And the test still passed because of "0 == 0".

This fix is to have more strict checking of the FDB entry
number. If any of the FDB count is 0, fail the test.

What's more, the reason of 0 FDB count is that config
reloading takes longer than before. The PTF script has
stopped running when FDB was cleared during config
reloading. The fix is to increase the PTF script running
time from 200 seconds to 300 seconds.

The third change is to replace raw config reloading
command with calling common_tasks/reload_config.yml

Signed-off-by: Xin Wang <xinw@mellanox.com>