[platform] optimize reboot-cause test case (#1167)
* [platform/test_reboot]

Only check reboot-cause for power off and watchdog reboot.
The power off and watchdog reboot share the same working flow with cold reboot except that their reboot causes differ and the motivation of them is to verify whether the reboot cause is correct. In this sense, it is unnecessary to do the full check for them.