[test behavior] disable allow_recover by default (#1241)
Allow_recover knob controls the behavior of test infrastructure. When set
to true, it allows a test to recover DUT in bad state before starting test.

With SONiC stability increases, we should by default disable recovery, and
surface more issues. If a test failed in the middle, it should be the test
itself to exit with a healthy DUT.

Signed-off-by: Ying Xie <ying.xie@microsoft.com>