[advanced-reboot] assert on total_disrupt_time
We missed a bug causing total_disrupt_time to be > 1 sec.
This is because the test asserts longest_disrupt_time < 1 sec.
However, there might be a lot of small periods of time which are < 1 sec
which result in total big downtime.
The fix is to assert also on total_disrupt_time in warm reboot test.
For fast reboot the test does not calculate total_disrupt_time because
it is assumed the lognest_disrupt_time == total_disrupt_time in case of
fast reboot.

Signed-off-by: Stepan Blyschak <stepanb@mellanox.com>