Fix issue of bad pfcwd docker image filename copied to fanout
The pfcwd docker image downloaded from a URL is saved to tmp folder
with a temporarily file name. Then it is copied to fanout and the
temporarily file name is kept. Subsequent "docker load" will complain
file not found and fail. This issue was not caught in previous testing
because we already have a pfcwd_storm.tgz file on fanout switch. The
"docker load" command was loading from the old image, not from the
just copied image.

The fix is to copy the temporarily downloaded file to fanout switch
with its original image name, not the temporarily filename.

Signed-off-by: Xin Wang <xinw@mellanox.com>