Add cli argument for skipping pytest sanity check
This change is an enhancement to the pytest sanity check. Previously, to skip
sanity check, we have to add pytest marker with "skip_sanity=True" in test
script. It is not convenient enough.

With this change, we can append argument "--skip_sanity" to the pytest cli
to skip pytest sanity check for test scripts.

Signed-off-by: Xin Wang <xinw@mellanox.com>