[platform fixtures] copy a few shared files/fixtures from tests/platform to tests/common/platform
This is the first PR of a serial of refactoring and clean up actions.

The intention / end goal is that for shared fixtured/functions, we
want to organize them into a nice folder structure for tests to use.

This move didn't deprecate/change any existing reference/usage of
these fixtures/functions in their original locations.

Moved files has been tested with a mock test that didn't go in with
this change.

Next steps would be further refactoring the existing code to move
more fixtures/functions/references over here. And eventually
deprecate the existing duplicate of these files.

Signed-off-by: Ying Xie <ying.xie@microsoft.com>