[advanced-reboot] The destination IP address of the packet chosen by the ansible script match static route of vlan1000 sometimes, which will result in the error.
One of dataplane traffic is sent from vlan interfaces to portchannel ports. The destination IP address of the packet is chosen randomly from valid choices. The found that destination IP address traffic chosen by the ansible script match static route of vlan1000 sometimes, which will result in the error.

Modify the algorithm to choose the destination IP from the given route entries in the fib database.
1.Traverse through all of the route entries in the fib and keep the route entries with the desired next hop in a list
2.Emits an error when the list created in the Step 1 is empty.
3.Choose a route entry randomly from the list created in the Step 1.
4.Choose a ip based on the route entry chosen in the Step 3.