Deprecate fixture testbed_devices
The testbed_devices fixture is not a good design. Multiple testbed device
objects are initialized in this fixture. It is not common that a test
script needs all the devices. This fixture may cause unnecessary overhead.
It would be better for test scripts to use the fixtures for different
devices on needed basis.

Changes:
1. Remove the definition of fixture testbed_devices.
2. Replace the call to testbed_devices fixture with other fixtures that
   fit better.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>