Announce vips routes for bgp_multipath_relax testing
Previously the vips routes were statically configured in the VMs for
bgp_multipath_relax testing. After the change of announcing
routes from exabgp in PTF to VMs, the vips routes were missing.
The bgp_multipath_relax testing depends on the vips routes failed.

This change is to announce the vips routes to VMs  that have the
vips routes configuration. The functions in the fib fixture definition
file were refactored a little bit.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>