[reload config] use /bin/bash as shell for config reload
This change was introduced in
https://github.com/Azure/sonic-mgmt/pull/1687 to fix connection reset
caused by 'config reload'. The fix was to redirect output to /dev/null.
Shell will interpret 'config reload -y &> /dev/null' as request to run
the command in the background instead of redirecting stdout/stderr to
/dev/null. It looks like not the redirection fixed the issue but the
fact that it was running in background. However, it introduced another
issue because config reload takes ~1.5m so that 90/120 sec wait time is
not enough time waiting for system to become ready.

Change-Id: I9d44c0a55e4fa016bbe177524891c9b14302c761
Signed-off-by: Stepan Blyschak <stepanb@mellanox.com>