[sonic image installer] remove config_db.json before booting into new image
When minigraph.xml exists, config_db.json can be generated from it.
In this case, before booting into an new image, remove the config_db.json
from /host/old_config to force the new image to load minigraph.

This is needed when nightly testbed is moving from an higer version to a
lower version.

Signed-off-by: Ying Xie <ying.xie@microsoft.com>