Add support of running general Ansible modules on EosHost
* Add support of running general Ansible modules on EosHost

The EOS hosts support two types of command line interface:
1. Linux shell
2. EOS shell

This change is to add the support of running general Ansible modules
that can only be executed under Linux shell.

* If not able to get the explicitly defined username and
password, fallback to use the value of "fanout_admin_user"
and "fanout_admin_password". It is unlikely to break the
existing code.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>