Improve the handling of VM kickstart failures
This commit improves the handling of VM kickstart failures

When we use the testbed-cli.sh tool to start a bunch of VMs,
occasionally some of the VMs failed to start during kickstart.
The current procedure is to destroy the VM and try to kickstart
it again. If the second try failed, then the whole starting
procedure will be stopped at the failed VM.

The testbed-cli.sh tool supports respin which is to destroy and
and start only the VMs specified in a list variable `respin_vms`.
We can re-run `testbed-cli.sh start-vms` and specify the VMs
that are not started in argument like `-e 'respin_vms=["VM0130","VM0131"]'.
Then the scripts will try to destroy the specified VMs and start
them again. The problem is that if the failed VM is in the front
part of the list, then we need to specify all the VMs starting
from the failed one in the list in argument `respin_vms`. This
would be tedious.

Changes in this commit:

* If kickstart some VMs failed, record the failed VM in a list
  and continue kickstarting rest of the VMs.
* Upon done, check if any VM has been recorded in the list.
  If yes, log the VMs that were not started and explicitly fail
  the playbook.
* Optionally we can manually destroy/undefine the failed VMs and
  delete their disk files.
* Then we can use the respin function to restart only the failed VMs.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>