Fix control plane acl testing for PR #1786
PR #1786 fixed an issue that no exception is raised for some failed ansible
modules. With that fix, the controle plane acl testing will fail. The reason is
that calling some of the ansible modules in this script is expected to fail.
However, because of the issue fixed by #1786, no exception is raised .
The test script explicitly check error message to determine if the ansible
module calling is failed or not.

With the fix of #1786, exception will be raised while calling ansible modules
that are expected to fail. We need to add `module_ignore_errors=True` for
these modules. Then we check the msg field of ansible module results to
determine whether the testing is expected or not.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>