[service_acl] Wait until SSH is stopped rather than waiting for start to timeout (#1820)
With the recent changes to caclmgrd, this check was not consistent, and could potentially fail with the following error:

```
TASK [test : Ensure the SSH port on the DuT becomes closed to us] ********************************************************************************************************************
Friday 26 June 2020  01:17:14 +0000 (0:00:05.323)       0:01:29.829 ***********
fatal: [sonic-dut-1]: FAILED! => {"msg": "The conditional check ''Timeout when waiting for search string OpenSSH' not in result.msg' failed. The error was: error while evaluating conditional ('Timeout when waiting for search string OpenSSH' not in result.msg): Unable to look up a name or access an attribute in template string ({% if 'Timeout when waiting for search string OpenSSH' not in result.msg %} True {% else %} False {% endif %}).\nMake sure your variable name does not contain invalid characters like '-': argument of type 'AnsibleUndefined' is not iterable"}
```

`result` could potentially be `AnsibleUndefined`.

This changes the logic to match that of the new pytest, which is a more appropriate method of checking that we are no longer able to SSH to the device, and no longer relies on parsing an error message (see https://github.com/Azure/sonic-mgmt/blob/master/tests/cacl/test_control_plane_acl.py#L34). Also remove unnecessary sleep, which also aligns more with the pytest version.