Fix unprintable RunAnsibleModuleFail exception (#1846)
Issue #1731 was not fully fixed by PR #1739. When RunAnsibleModuleFail exception
is raised by running ptf script, the unprintable exception issue was still observed.

The reason is that the ansible result may contain unicode string. Concatenating ascii
string with unicode string may raise uncaught exception and cause this issue.

The fix is to concatenate unicode string with the ansible result. Then encode it
to ascii string.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>