Add `--show-capture no` argument in `run_tests.sh` (#1935)
The run_tests.sh is able to store all the testing logs in
log files. It's unnecessary to show captured logs in
cli log. Otherwise, the cli log is also huge.
This commit add the `--show-capture no` argument
to the PYTEST_COMMON_OPTS in `run_tests.sh` to
disable that.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>