Use paramiko for ssh connection in test_wr_arp
The test_wr_arp script put ssh key of ptfhost on dut to have password less
ssh connection. Then the ptfhost can execute command remote on dut
through this ssh connection. The problem is that the ssh key saved in
~/.ssh/authorized_keys on dut host may not persist after warm reboot.
It would also be a problem when secure boot feature is introduced. This
PR updated the test_wr_arp script to use paramiko for ssh connection
and remote command execution between ptfhost and dut.

Changes:
* Update the ptf test script to use paramiko for SSH connection with dut
* Disabled loganalyzer because test_wr_arp need to do warm-reboot on dut

Signed-off-by: Xin Wang <xiwang5@microsoft.com>