[copp] Fix recent COPP test failures (#2030)
The COPP tests can occasionally trigger crashes and inconsistent behavior in orchagent. This is because restart swss does not provide strong guarantees for the ordering that services are restarted in. To mitigate this risk, we can use config reload as this ensures that services will be restarted in the proper order.

In addition, there was a recent behavioral change for T1 devices that causes DHCP packets to not be trapped to CPU. As a result, this case is now expected to fail.

Signed-off-by: Danny Allen <daall@microsoft.com>