[docker-setup]: Fix docker group membership check (#2070)
Prevent docker commands from running without first validating that user is a member of the docker group (by checking the exit code of `docker info`). 

It is possible for a user to have docker permissions without explicitly being in the docker group (shown by id -Gn <user> or groups <user>). The previous approach grepped the output of id -Gn for the docker group, which is not always reliable.