Fix issue of parsing yaml inventory file in the testbed_vm_info module (#2091)
If the veos inventory file is converted to yaml format, the testbed_vm_info module will be affected. Consequently `testbed-cli.sh gen-mg` and `testbed-cli.sh deploy-mg` that are dependent on this module will fail.

The reason is that `testbed_vm_info` parses the `veos` inventory file as plain text and expects the `veos` file to have the `ansible_host=x.x.x.x` variable on the same line with host name like `VM0001`.

Update the module to use the ansible `InventoryManager` for parsing the `veos` inventory file. Then it will always work no matter the inventory file is ini format or yaml format.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>