Fix the false alarm of checking reboot result
After PR "https://github.com/Azure/sonic-mgmt/pull/2063 Fix the issue of no 'failed' field in failed ansible module result "
is merged, the ansible module result will no longer have the 'failed' field filtered out. We now can count on the 'is_failed'
property of ansible module result to tell if it is failed or not. However, just assert that the result is failed if the 'failed' field
is in the result dict is not safe now. It's because the 'failed' field may have value 'false'. The reboot function in reboot.py
may raise false alarm "DUT was not shutdown".

The fix:
* Explicitly check the 'is_failed' property of localhost.wait_for result to determine whether the DUT is down or up as expected.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>