Add "router_mac" to SonicHost.facts
For PTF related tests, we often need to get the router MAC of the SONiC
switch and pass it to the PTF script in arguments. Many test scripts use
the 'setup' Ansible module to get the router MAC. However, the 'setup'
module gathers too much information and generats too much log. It is
a over kill to just get router MAC using the 'setup' module.

This enhancement gets the router MAC from running config and stores
it in the SonicHost.facts property. In test script, we can get it using
code like 'duthost.facts["router_mac"]'.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>