Wait for telemetry port up after telemetry service is restarted
This change is to fix the test_telemetry_ouput case failure. After the telemetry
service is restarted, the telemetry TCP port still needs some time to go up.
If immediately run the py_gnmicli.py script to get data from the telemetry
TCP port, the script may fail with "Connection refused". The fix added code
to wait for the telemetry TCP port. When the port is up, then proceed with
testing.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>