Remove unnecessary hard coded skip scripts in run_tests.sh
Two scripts "test_announce_routes.py" and "test_nbr_health.py" are hard coded
as scripts to be skipped in run_tests.sh. This caused these two scripts are not
executed while preparing the DUT for testing.

Indeed, either "-t TOPOLOGY" or "-c TEST_CASES" must be specified in
the run_tests.sh command line arguments. When "-t TOPOLOGY" is specified,
the scripts with "utils" mark will be skipped during test execution. It's
unnecessary to specify them in the skip list. When "-c TEST_CASE" is specified,
the person running the command is responsible for the correctness of the
TEST_CASES list.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>