Refactor: rename the 'testbed' fixture to 'tbinfo'
In the future, we need to support multiple dut and chassis testing.
One possible solution is to add a 'testbed' fixture represents all the
DUT nodes. When the testbed has only single DUT, the behavior
of testbed object is same as duthost. When there are multiple DUT
nodes, the testbed object can encapsulate code to handle the
complexity of dealing with multiple DUT nodes that may have
different roles. Currently the 'testbed' name is taken for providing
testbed information to test cases. We can't think of a better name
for the object that will be repensending multiple DUT nodes in a
testbed. So, this PR is to refactor the code to rename the existing
'testbed' fixture to 'tbinfo' which is more meaningful.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>