Update hook for custom fixture.
Fixture check_dut_asic_type is set to function level, as a result, it
always runs after all module-level fixtures. This commit update the hook
for custom fixture. The check_dut_asic_type fixture is inserted into
fixturelist before all module-level fixture, so that these fixtures
no longer runs if the case is to be skipped.