Fix conn_graph_facts issues of `anchor` and `filename` options (#2249)
* `anchor` option issue
When use `anchor` to get the whole connection graph facts for configuring root
fanout switch, the `host` and `hosts` options are not given. So the `hostnames`
variable will have value [None] which will not be evaluated to False. Then the
`target` variable will take the value of `hostnames`, i.e. [None]. Eventually the empty
graph file will be used.

* `filename` option issue
When the `filename` argument is specified, need to open the specified file under
path LAB_GRAPHFILE_PATH.

Changes:
* Fix the issue of parsing wrong connection graph file when the `anchor` option
  is used.
* Fix the issue of unable to find connection graph file when the `filename`
  option is used.
* Set host, hosts and anchor mutually exclusive.
* Update the documentation.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>