Fix test_show_vlan_config
An interface is added to Vlan100 to test. However, the test script
remove Vlan100 at the end of test without removing added interface. As a
result, a WARNING log is printed 'ERR swss#orchagent: :- removeVlan:
Failed to remove non-empty VLAN Vlan100' and causes test error. This
commit address the issue.