Merged PR 3583648: Fixing the incorrect PTF IP allocation for vms2-4-t0-2700
The IP assigned for testbed ` vms2-4-t0-2700` in testbed.csv and `str` inventory are different. Moreover, these IPs are not owned by this testbed on the OneNote page for IP allocation.

Picking a new IP and file changes fixed the AnsibleConnection Failures seen in the recent Pytests.

Failure that is being fixed:
https://sonic-jenkins.corp.microsoft.com:3389/job/Testbed/job/announce-routes/8/console
Pass log:
https://sonic-jenkins.corp.microsoft.com:3389/job/Testbed/job/announce-routes/9/console