[test_crm] Fix test_acl_counter and test_crm_nexthop_group in test_crm (#2272)
* Fix test_acl_counter

Because the acl counter is relatively large in some DUT, and even 1% of
the acl counter will be larger than the avaliable acl entry. This PR
adds a limit to the acl entry that are applied to DUT.

* Update log for skip check

* Fix issue in test_crm_nexthop

The test case will attempt to insert more nexthop members than the total
available number on some devices. As a result, SAI will return error.
This commit add a check and get the min value between avaiable nexthop_group_member and
nexthop_group_object before inserting new nexthops.