Use new shell_cmds module to enable/disable container autorestart (#2263)
* Use new shell_cmds module to enable/disable container autorestart

The ansible built-in modules like `command` only supports running single command.
The `shell` module can run multiple commands by concatenate the commands using
` && `. But the problem is that all the commands output are concatenated too. It's
hard to tell which output belongs to which command. Because of this limitation,
the fixture `disable_container_autorestart` run commands one by one to disable
or enable container autorestart for each containers. As an autoused fixture, it
caused too much unnecessary overhead.

This commit added a new customized ansible module `shell_cmds`. It accepts
a list of commands in string and runs the commands in sequential on remote
host. Output of this module contains a list of results. One result for each
command. The `disable_container_autorestart` fixture is also updated to use
this new module. Tested on vsimage, using this new `shell_cmds` module
can reduce around 8 seconds of overhead caused by fixture
`disable_container_autorestart`

Signed-off-by: Xin Wang <xiwang5@microsoft.com>

* Add option continue_on_fail for shell_cmds module