[pfcwd] Filter rx ports out of test ports
If port `A` is used as test port and rx port for port `B`, and the test
conducts `storm_restore_path` over port `B` prior to port `A`. The
traffic verification for port `B` will fail because it uses port `A` as
rx port, which is currently in stormed state.

So filter out those ports that act as rx port out of selected test
ports.

Signed-off-by: Longxiang Lyu <lolv@microsoft.com>