Improve wait_dut and prepare_ptf for the vxlan decap test
The logic of wait_dut in the vxlan decap PTF test script is complicated and
incorrect. There is no sleep in each iteration. The maximum loop is limited
by number of `timeout`, not by `timeout` seconds.

This change simplified the wait_dut function and limit the maximum loop time
to `timeout` seconds. Another change is to deprecate the call `setup` ansible
module on duthost and ptfhost. The ptf_facts returned by ptfhost.setup() is
unused and can be removed. The dut_facts returned by duthost.setup() is for
getting router MAC. There is an easier to get router MAC by
duthost.facts["router_mac"]

Signed-off-by: Xin Wang <xiwang5@microsoft.com>