[Multi asic]:parameterize asic_index and dut_index (#2245)
In case of SONiC on multi asic platforms, each asic is bound to a namespace. The dataplane and control plane containers are replicated per namespace. The tests which are testing functionality within a namespace, can be repeated on all namespaces by passing asic_index as parameter to the test. This way the testing logic remains. The test will be repeated N times, where the N is the num_asics.

Recently sonic-mgmt introduced the support for multiple duts in the topology. We are proposing to pass dut_index as parameter to the testcase. So, the test will be repeated on each dut in the topology.

If asic_index and dut_index is used in the test, the test will be repeated on each asic of each dut in the topology.

Signed-off-by: Arvindsrinivasan Lakshminarasimhan <arlakshm@microsoft.com>