Stabilize the NTP test (#2426)
Script test_ntp.py is failing and caused sonic-mgmt PR testing failed.
This change is to stabilizie the ntp test to ensure that the NTP server
on PTF is always restarted to take the configuration change in setup.

This change will force the ntpd on PTF to always use local time. Then 
the ntp server on PTF docker will always be "synchronized". This test 
will be more stable. Since we are not updating time of the PTF container, 
time of PTF will be same as host. Eventually, time of DUT will be kind 
of synchronized with host.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>